[cpu] Replace _xgetbv identifier with xgetbv
Replace use of the reserved _xgetbv identifier with xgetbv. Clang built from trunk now implements _xgetbv, causing errors as it is now redefined.

Modeled after https://crrev.com/c/1414858 which is a recent fix to v8 (which this appears to be a fork of), although this fork does not use the msvc fallbacks.